chore: move type definitions into types/base to update docs
