+++ packages/sdk-browser/src/__types__/base.ts
@@ -3,6 +3,7 @@ import type { WalletInterface, WalletStamper } from "@turnkey/wallet-stamper";
 import type * as SdkApiTypes from "../__generated__/sdk_api_types";
 import type { WebauthnStamper } from "@turnkey/webauthn-stamper";
 import type { IframeStamper } from "@turnkey/iframe-stamper";
+import type { TurnkeyIframeClient } from "../sdk-client";
 
 export type GrpcStatus = {
   message: string;
@@ -19,6 +20,88 @@ export enum MethodType {
 export enum SessionType {
   READ_ONLY = "SESSION_TYPE_READ_ONLY",
   READ_WRITE = "SESSION_TYPE_READ_WRITE",
+}
+
+export type Session = {
+  sessionType: SessionType;
+  userId: string;
+  organizationId: string;
+  expiry: number; // Unix timestamp representing the expiry of the session set by the server
+  token: string; // credentialBundle (read-write) or read token
+};
+
+/**
+ * This interface defines the structure of user data that will be stored in local storage
+ * when using one of the login methods.
+ */
+export interface User {
+  // Unique identifier for the user.
+  userId: string;
+
+  // Username of the user.
+  username: string;
+
+  // Organization details associated with the user.
+  organization: {
+    // Unique identifier for the organization.
+    organizationId: string;
+
+    // Name of the organization.
+    organizationName: string;
+  };
+
+  // Session information for the user, which can be either read-only or read-write.
+  session:
+    | {
+        // Optional read-only session details.
+        read?: ReadOnlySession;
+
+        // Optional read-write session details.
+        write?: ReadWriteSession;
+
+        // Authenticated client associated with the session.
+        authClient: AuthClient;
+      }
+    | undefined;
+}
+
+/**
+ * Represents an expiring read-only session with a token for read-only requests.
+ */
+export interface ReadOnlySession {
+  /**
+   * This token is generated by the `createReadOnlySession` method in the `TurnkeySDKClientBase` class.
+   * It is used in the `X-Session` header for read-only requests to Turnkey's API.
+   *
+   * @see {@link TurnkeySDKClientBase#createReadOnlySession}
+   */
+  token: string;
+
+  /**
+   * The expiry time of the token in seconds since the Unix epoch.
+   * This indicates when the token will no longer be valid.
+   */
+  expiry: number;
+}
+
+/**
+ * Represents an expiring read-write session with a credential bundle for read-write requests.
+ */
+export interface ReadWriteSession {
+  /**
+   * This bundle is generated by the `createReadWriteSession` and `createApiKeys` methods in the `TurnkeySDKClientBase` and `TurnkeyPasskeyClient` classes.
+   * It is injected into the iFrame client to enable read and write operations.
+   *
+   * @see {@link TurnkeySDKClientBase#createReadWriteSession}
+   * @see {@link TurnkeyPasskeyClient#createApiKeys}
+   */
+  credentialBundle: string;
+
+  /**
+   * The expiry time of the credential bundle in seconds since the Unix epoch.
+   * This indicates when the credential bundle will no longer be valid.
+   */
+  expiry: number;
 }
 
 export type TStamp = {
@@ -125,6 +208,31 @@ export interface PasskeyClientParams {
   timeout?: number;
   userVerification?: UserVerificationRequirement;
   allowCredentials?: PublicKeyCredentialDescriptor[];
+}
+
+export interface RefreshSessionParams {
+  sessionType: SessionType;
+  expirationSeconds?: string | undefined;
+  targetPublicKey?: string;
+}
+
+export interface LoginWithBundleParams {
+  bundle: string;
+  expirationSeconds?: string;
+}
+
+export interface LoginWithPasskeyParams {
+  sessionType: SessionType;
+  iframeClient: TurnkeyIframeClient;
+  expirationSeconds?: string | undefined;
+  targetPublicKey?: string;
+}
+
+export interface LoginWithWalletParams {
+  sessionType: SessionType;
+  iframeClient: TurnkeyIframeClient;
+  expirationSeconds?: string | undefined;
+  targetPublicKey?: string;
 }
 
 export interface TurnkeyWalletClientConfig extends SDKClientConfigWithStamper {
