+++ packages/sdk-browser/src/models.ts
@@ -1,79 +1,3 @@
-import type { AuthClient } from "@types";
-
-/**
- * This interface defines the structure of user data that will be stored in local storage
- * when using one of the login methods.
- */
-export interface User {
-  // Unique identifier for the user.
-  userId: string;
-
-  // Username of the user.
-  username: string;
-
-  // Organization details associated with the user.
-  organization: {
-    // Unique identifier for the organization.
-    organizationId: string;
-
-    // Name of the organization.
-    organizationName: string;
-  };
-
-  // Session information for the user, which can be either read-only or read-write.
-  session:
-    | {
-        // Optional read-only session details.
-        read?: ReadOnlySession;
-
-        // Optional read-write session details.
-        write?: ReadWriteSession;
-
-        // Authenticated client associated with the session.
-        authClient: AuthClient;
-      }
-    | undefined;
-}
-
-/**
- * Represents an expiring read-only session with a token for read-only requests.
- */
-export interface ReadOnlySession {
-  /**
-   * This token is generated by the `createReadOnlySession` method in the `TurnkeySDKClientBase` class.
-   * It is used in the `X-Session` header for read-only requests to Turnkey's API.
-   *
-   * @see {@link TurnkeySDKClientBase#createReadOnlySession}
-   */
-  token: string;
-
-  /**
-   * The expiry time of the token in seconds since the Unix epoch.
-   * This indicates when the token will no longer be valid.
-   */
-  expiry: number;
-}
-
-/**
- * Represents an expiring read-write session with a credential bundle for read-write requests.
- */
-export interface ReadWriteSession {
-  /**
-   * This bundle is generated by the `createReadWriteSession` and `createApiKeys` methods in the `TurnkeySDKClientBase` and `TurnkeyPasskeyClient` classes.
-   * It is injected into the iFrame client to enable read and write operations.
-   *
-   * @see {@link TurnkeySDKClientBase#createReadWriteSession}
-   * @see {@link TurnkeyPasskeyClient#createApiKeys}
-   */
-  credentialBundle: string;
-
-  /**
-   * The expiry time of the credential bundle in seconds since the Unix epoch.
-   * This indicates when the credential bundle will no longer be valid.
-   */
-  expiry: number;
-}
-
 export interface SubOrganization {
   organizationId: string;
   organizationName: string;
