+++ packages/sdk-server/src/turnkey-helpers.ts
@@ -1,48 +1,8 @@
-export interface WalletAccount {
-  curve: "CURVE_SECP256K1" | "CURVE_ED25519";
-  pathFormat: "PATH_FORMAT_BIP32";
-  path: string;
-  addressFormat:
-    | "ADDRESS_FORMAT_ETHEREUM"
-    | "ADDRESS_FORMAT_UNCOMPRESSED"
-    | "ADDRESS_FORMAT_COMPRESSED"
-    | "ADDRESS_FORMAT_SOLANA"
-    | "ADDRESS_FORMAT_COSMOS"
-    | "ADDRESS_FORMAT_TRON"
-    | "ADDRESS_FORMAT_SEI"
-    | "ADDRESS_FORMAT_XLM"
-    | "ADDRESS_FORMAT_BITCOIN_MAINNET_P2PKH"
-    | "ADDRESS_FORMAT_BITCOIN_MAINNET_P2WPKH"
-    | "ADDRESS_FORMAT_BITCOIN_MAINNET_P2WSH"
-    | "ADDRESS_FORMAT_BITCOIN_MAINNET_P2TR"
-    | "ADDRESS_FORMAT_BITCOIN_MAINNET_P2SH"
-    | "ADDRESS_FORMAT_BITCOIN_TESTNET_P2PKH"
-    | "ADDRESS_FORMAT_BITCOIN_TESTNET_P2WPKH"
-    | "ADDRESS_FORMAT_BITCOIN_TESTNET_P2WSH"
-    | "ADDRESS_FORMAT_BITCOIN_TESTNET_P2TR"
-    | "ADDRESS_FORMAT_BITCOIN_TESTNET_P2SH"
-    | "ADDRESS_FORMAT_BITCOIN_SIGNET_P2PKH"
-    | "ADDRESS_FORMAT_BITCOIN_SIGNET_P2WPKH"
-    | "ADDRESS_FORMAT_BITCOIN_SIGNET_P2WSH"
-    | "ADDRESS_FORMAT_BITCOIN_SIGNET_P2TR"
-    | "ADDRESS_FORMAT_BITCOIN_SIGNET_P2SH"
-    | "ADDRESS_FORMAT_BITCOIN_REGTEST_P2PKH"
-    | "ADDRESS_FORMAT_BITCOIN_REGTEST_P2WPKH"
-    | "ADDRESS_FORMAT_BITCOIN_REGTEST_P2WSH"
-    | "ADDRESS_FORMAT_BITCOIN_REGTEST_P2TR"
-    | "ADDRESS_FORMAT_BITCOIN_REGTEST_P2SH"
-    | "ADDRESS_FORMAT_DOGE_MAINNET"
-    | "ADDRESS_FORMAT_DOGE_TESTNET"
-    | "ADDRESS_FORMAT_SUI"
-    | "ADDRESS_FORMAT_APTOS"
-    | "ADDRESS_FORMAT_XRP"
-    | "ADDRESS_FORMAT_TON_V3R2"
-    | "ADDRESS_FORMAT_TON_V4R2";
-}
-
 // ----------------------------
 // CURVE_SECP256K1 Accounts
 // ----------------------------
+
+import type { WalletAccount } from "./__types__/base";
 
 // Ethereum
 export const defaultEthereumAccountAtIndex = (
